+++ test/Table/TableStyleSpec.ts
@@ -51,6 +51,61 @@ describe("Table Style", function() {
       expect(activeStyle.tableColorMap.binColors.length).toEqual(7);
       expect(activeStyle.tableColorMap.binMaximums.length).toEqual(7);
       expect(activeStyle.colorMap instanceof DiscreteColorMap).toBeTruthy();
+
+      expect(
+        (activeStyle.colorMap as DiscreteColorMap).colors.map(c =>
+          c.toCssHexString()
+        )
+      ).toEqual([
+        "#fee5d9",
+        "#fcbba1",
+        "#fc9272",
+        "#fb6a4a",
+        "#ef3b2c",
+        "#cb181d",
+        "#99000d"
+      ]);
+    });
+
+    it(" - uses DiscreteColorMap if set binMaximums", async function() {
+      csvItem.setTrait(
+        "definition",
+        "url",
+        "/test/csv/youth-unemployment-rate-2018.csv"
+      );
+      csvItem.setTrait("definition", "styles", [
+        createStratumInstance(TableStyleTraits, {
+          id: "youth unemployment (%)",
+          color: createStratumInstance(TableColorStyleTraits, {
+            binMaximums: [8, 10, 15, 20, 30, 50],
+            colorPalette: "PiYG"
+          })
+        })
+      ]);
+      await csvItem.loadMapItems();
+
+      const activeStyle = csvItem.activeTableStyle;
+      const colorColumn = activeStyle.colorColumn;
+      expect(colorColumn).toBeDefined();
+      expect(colorColumn!.type).toBe(4);
+      expect(colorColumn!.values.length).toBe(86);
+
+      expect(activeStyle.tableColorMap.binColors.length).toEqual(6);
+      expect(activeStyle.tableColorMap.binMaximums.length).toEqual(6);
+      expect(activeStyle.colorMap instanceof DiscreteColorMap).toBeTruthy();
+
+      expect(
+        (activeStyle.colorMap as DiscreteColorMap).colors.map(c =>
+          c.toCssHexString()
+        )
+      ).toEqual([
+        "#c51b7d",
+        "#e9a3c9",
+        "#fde0ef",
+        "#e6f5d0",
+        "#a1d76a",
+        "#4d9221"
+      ]);
     });
 
     it(" - uses ContinuousColorMap by default", async function() {
@@ -88,6 +143,18 @@ describe("Table Style", function() {
 
       expect(activeStyle.colorMap instanceof EnumColorMap).toBeTruthy();
       expect((activeStyle.colorMap as EnumColorMap).colors.length).toBe(6);
+      expect(
+        (activeStyle.colorMap as EnumColorMap).colors.map(c =>
+          c.toCssHexString()
+        )
+      ).toEqual([
+        "#f2f3f4",
+        "#ffb300",
+        "#803e75",
+        "#ff6800",
+        "#a6bdd7",
+        "#c10020"
+      ]);
     });
   });
 });