+++ CHANGES.md
@@ -54,7 +54,7 @@ Change Log
   * Replaced with `disableZoomTo` in `MappableTraits`
 * Clean up `showsInfo`
   * Replaced with `disableAboutData` in `CatalogMemberTraits`
-* Fix `DiscreteColorMap` bug with `binColors`
+* Fix `DiscreteColorMap` bug with `binColors` and added warning message if `colorPalette` is invalid.
 * [The next improvement]
 
 #### 8.0.0-alpha.87